refactor: orderLineItem이 menu를 간접참조하고, 이름과 가격을 가지도록 변경
