+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -5,6 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import kitchenpos.vo.Name;
 
 @Entity
 public class MenuGroup {
