+++ src/main/java/kitchenpos/menu/domain/Product.java
@@ -7,6 +7,8 @@ import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import kitchenpos.vo.Name;
+import kitchenpos.vo.Price;
 
 @Entity
 public class Product {
