+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -21,9 +21,6 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    //    @ManyToOne(fetch = FetchType.LAZY)
-//    @JoinColumn(nullable = false)
-//    private OrderTable orderTable;
     private Long orderTableId;
 
     private LocalDateTime orderedTime;
