+++ src/main/resources/db/migration/V3__Update_order_line_item_column.sql
@@ -1,0 +1,4 @@
+ALTER TABLE order_line_item
+    ADD COLUMN name varchar(255) not null;
+ALTER TABLE order_line_item
+    ADD COLUMN price bigint not null;
