+++ src/test/java/kitchenpos/menu/domain/PriceTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
+import kitchenpos.vo.Price;
 import org.junit.jupiter.api.Test;
 
 class PriceTest {
